+++ agentloop-20260520/src/Models/GetPipelineResponseBody/sink/condition/routes/sink.php
@@ -2,19 +2,13 @@
 
 // This file is auto-generated, don't edit it. Thanks.
 
-namespace AlibabaCloud\SDK\AgentLoop\V20260520\Models\CreatePipelineRequest;
+namespace AlibabaCloud\SDK\AgentLoop\V20260520\Models\GetPipelineResponseBody\sink\condition\routes;
 
 use AlibabaCloud\Dara\Model;
-use AlibabaCloud\SDK\AgentLoop\V20260520\Models\CreatePipelineRequest\sink\condition;
-use AlibabaCloud\SDK\AgentLoop\V20260520\Models\CreatePipelineRequest\sink\dataset;
+use AlibabaCloud\SDK\AgentLoop\V20260520\Models\GetPipelineResponseBody\sink\condition\routes\sink\dataset;
 
 class sink extends Model
 {
-    /**
-     * @var condition
-     */
-    public $condition;
-
     /**
      * @var dataset
      */
@@ -25,16 +19,12 @@ class sink extends Model
      */
     public $type;
     protected $_name = [
-        'condition' => 'condition',
         'dataset' => 'dataset',
         'type' => 'type',
     ];
 
     public function validate()
     {
-        if (null !== $this->condition) {
-            $this->condition->validate();
-        }
         if (null !== $this->dataset) {
             $this->dataset->validate();
         }
@@ -44,10 +34,6 @@ class sink extends Model
     public function toArray($noStream = false)
     {
         $res = [];
-        if (null !== $this->condition) {
-            $res['condition'] = null !== $this->condition ? $this->condition->toArray($noStream) : $this->condition;
-        }
-
         if (null !== $this->dataset) {
             $res['dataset'] = null !== $this->dataset ? $this->dataset->toArray($noStream) : $this->dataset;
         }
@@ -67,10 +53,6 @@ class sink extends Model
     public static function fromMap($map = [])
     {
         $model = new self();
-        if (isset($map['condition'])) {
-            $model->condition = condition::fromMap($map['condition']);
-        }
-
         if (isset($map['dataset'])) {
             $model->dataset = dataset::fromMap($map['dataset']);
         }